+++ app/Models/BarIngredient.php
@@ -1,0 +1,33 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Kami\Cocktail\Models;
+
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+
+class BarIngredient extends Model
+{
+    /** @use \Illuminate\Database\Eloquent\Factories\HasFactory<\Database\Factories\BarIngredientFactory> */
+    use HasFactory;
+
+    public $timestamps = false;
+
+    /**
+     * @return BelongsTo<Ingredient, $this>
+     */
+    public function ingredient(): BelongsTo
+    {
+        return $this->belongsTo(Ingredient::class);
+    }
+
+    /**
+     * @return BelongsTo<Bar, $this>
+     */
+    public function bar(): BelongsTo
+    {
+        return $this->belongsTo(Bar::class);
+    }
+}