+++ app/Models/Export.php
@@ -12,9 +12,12 @@ use Illuminate\Support\Facades\Storage;
 use Kami\Cocktail\Models\Concerns\HasAuthors;
 use Kami\Cocktail\Models\Concerns\HasBarAwareScope;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 
 class Export extends Model
 {
+    /** @use \Illuminate\Database\Eloquent\Factories\HasFactory<\Database\Factories\ExportFactory> */
+    use HasFactory;
     use HasAuthors;
     use HasBarAwareScope;