+++ database/migrations/2024_11_09_133632_migrate_bar_shelf_data.php
@@ -1,0 +1,27 @@
+<?php
+
+use Illuminate\Support\Facades\DB;
+use Illuminate\Database\Migrations\Migration;
+
+return new class () extends Migration {
+    /**
+     * Run the migrations.
+     */
+    public function up(): void
+    {
+        DB::statement(<<<SQL
+            INSERT INTO bar_ingredients (bar_id, ingredient_id) 
+            SELECT bar_id, ingredient_id FROM user_ingredients 
+                JOIN bar_memberships ON user_ingredients.bar_membership_id = bar_memberships.id 
+            WHERE user_id = (SELECT created_user_id FROM bars)
+        SQL);
+    }
+
+    /**
+     * Reverse the migrations.
+     */
+    public function down(): void
+    {
+        DB::statement('DELETE FROM bar_ingredients');
+    }
+};